requirements.txt
pymongo
croniter
kubernetes
python-dateutil